+++ example.js
@@ -1,12 +1,14 @@
 const { start, color, createMap } = require("./index");
 
+const { bitmaps } = require("./exported-animation.json");
+
 // prettier-ignore
 const map = createMap([
   'cccccccccccccccc',
   'c              c',
   'c              c',
-  'c    56789     c',
-  'c    abcde     c',
+  'c              c',
+  'c              c',
   'c              c',
   'c              c',
   'cccccccccccccccc',
@@ -34,8 +36,9 @@ start({
   },
 
   draw: (frame, { x, y }) => {
-    frame.mset(0, 0, map);
-    frame.rect(3, 1, 12, 7, color.pink);
+    frame.bset(bitmaps[0]);
+    frame.mset(0, 0, map, 0);
+    frame.rect(1, 1, 15, 7, color.pink);
     frame.pset(x, y, color.red);
   }
 });